add Promise to reset
Hello, if a client switches from one account to another, the chatwood stops working. We need to wait for the restart to complete everywhere. I added a promise, and that's enough for me, but I think you might need to modify the code on your end that uses the restart.